requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.2.3
openpyxl==3.1.5
python-docx==1.1.2
google-genai==1.19.0
gunicorn==20.1.0
